Bug Fixes & Improved Sample
