+++ add-remote.sh
@@ -1,0 +1,2 @@
+git remote add $1 https://github.com/$1/showcaseview
+git fetch $1
