+++ sample/build.gradle
@@ -36,9 +36,9 @@ apply plugin: 'com.android.application'
 dependencies {
     compile project(':library')
     //compile 'com.github.amlcurran.showcaseview:library:5.0.0-SNAPSHOT'
-    compile 'com.android.support:support-v4:21.0.0'
+    compile 'com.android.support:support-v4:21.0.3'
     compile 'com.nineoldandroids:library:2.4.0'
-    compile 'com.android.support:appcompat-v7:21.0.0'
+    compile 'com.android.support:appcompat-v7:21.0.3'
 }
 
 android {
