+++ sample/src/main/AndroidManifest.xml
@@ -33,14 +33,13 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".v14.ActionItemsSampleActivity"
-            android:theme="@style/SampleTheme" />
         <!--<activity-->
         <!--android:name=".v14.MultipleActionItemsSampleActivity"-->
         <!--android:theme="@style/SampleTheme"/>-->
-        <!--<activity-->
-        <!--android:name=".legacy.MultipleShowcaseSampleActivity"/>-->
+
+        <activity
+            android:name=".MultipleShowcaseSampleActivity"/>
+
         <activity android:name=".animations.AnimationSampleActivity" />
 
         <activity android:name=".SingleShotActivity" />
