+++ sample/src/main/java/com/github/amlcurran/showcaseview/sample/MemoryManagementTesting.java
@@ -24,7 +24,7 @@ public class MemoryManagementTesting extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.main);
+        setContentView(R.layout.activity_sample);
 
 //        ShowcaseViews showcaseViews = new ShowcaseViews(this);
 //
