+++ sample/src/main/res/layout/activity_sample_multiple_showcase.xml
